feat: ヘッダーに記載されていた諸関数の実装をcppファイルに移動
